fix: исправление замечаний
