+++ src/main/java/ru/practicum/shareit/booking/BookingServiceImpl.java
@@ -7,16 +7,14 @@ import ru.practicum.shareit.booking.enums.Status;
 import ru.practicum.shareit.exceptions.NotFoundException;
 import ru.practicum.shareit.exceptions.UnknownStateException;
 import ru.practicum.shareit.exceptions.ValidationException;
-import ru.practicum.shareit.item.ItemRepository;
-import ru.practicum.shareit.item.ItemValidationService;
 import ru.practicum.shareit.item.Item;
-import ru.practicum.shareit.user.UserRepository;
-import ru.practicum.shareit.user.UserValidationService;
+import ru.practicum.shareit.item.ItemRepository;
 import ru.practicum.shareit.user.User;
+import ru.practicum.shareit.user.UserRepository;
+import ru.practicum.shareit.util.BookingValidationService;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -25,20 +23,18 @@ public class BookingServiceImpl implements BookingService {
     private final BookingRepository bookingRepository;
     private final UserRepository userRepository;
     private final ItemRepository itemRepository;
-    private final BookingValidationService bookingValidationService;
-    private final UserValidationService userValidationService;
-    private final ItemValidationService itemValidationService;
 
     @Override
     public Booking addBooking(BookingDto bookingDto, long bookerId) {
-        bookingValidationService.validate(bookingDto);
-        userValidationService.checkUserIsExist(bookerId);
-        itemValidationService.checkItemIsExist(bookingDto.getItemId());
+        BookingValidationService.validate(bookingDto);
+        Item item = itemRepository.findById(bookingDto.getItemId()).orElseThrow(
+                () -> new NotFoundException("Item id: " + bookingDto.getItemId() + " not found")
+        );
+        User booker = userRepository.findById(bookerId).orElseThrow(
+                () -> new NotFoundException("User id: " + bookerId + " not found"));
 
-        Item item = itemRepository.findById(bookingDto.getItemId()).get();
         if (!item.getAvailable())
             throw new ValidationException("Item is not available");
-        User booker = userRepository.findById(bookerId).get();
 
         if (booker.getId() == item.getOwner().getId())
             throw new NotFoundException("Owner cannot book his item");
@@ -49,14 +45,16 @@ public class BookingServiceImpl implements BookingService {
 
     @Override
     public Booking setOwnersDecision(long bookingId, boolean approved, long ownerId) {
-        bookingValidationService.checkBookingIsExist(bookingId);
-        userValidationService.checkUserIsExist(ownerId);
+        Booking booking = bookingRepository.findById(bookingId).orElseThrow(
+                () -> new NotFoundException("Booking id " + bookingId + " not found")
+        );
+        User owner = userRepository.findById(ownerId).orElseThrow(
+                () -> new NotFoundException("User id " + ownerId + " not found")
+        );
 
-        Booking booking = bookingRepository.findById(bookingId).get();
         if (booking.getStatus() == Status.APPROVED) {
             throw new ValidationException("Booking is already approved");
         }
-        User owner = userRepository.findById(ownerId).get();
 
         if (owner.getId() != booking.getItem().getOwner().getId())
             throw new NotFoundException("User is not owner by item");
@@ -71,26 +69,25 @@ public class BookingServiceImpl implements BookingService {
 
     @Override
     public Booking findBookingById(long bookingId, long userId) {
-        Optional<Booking> bookingOpt = bookingRepository.findById(bookingId);
-        if (bookingOpt.isEmpty()) {
-            throw new NotFoundException("Booking not found");
-        } else {
-            Booking booking = bookingOpt.get();
+        Booking booking = bookingRepository.findById(bookingId).orElseThrow(
+                () -> new NotFoundException("Booking id: " + bookingId + " not found")
+        );
 
-            if (booking.getBooker().getId() == userId ||
-            booking.getItem().getOwner().getId() == userId) {
-                return booking;
-            } else {
-                throw new NotFoundException("User do not have access to this booking");
-            }
+        if (booking.getBooker().getId() == userId ||
+        booking.getItem().getOwner().getId() == userId) {
+            return booking;
+        } else {
+            throw new NotFoundException("User do not have access to this booking");
         }
+
     }
 
     @Override
     public List<Booking> findAllByBooker(long userId, String state) {
         LocalDateTime now = LocalDateTime.now();
-        userValidationService.checkUserIsExist(userId);
-        User user = userRepository.findById(userId).get();
+        User user = userRepository.findById(userId).orElseThrow(
+                () -> new NotFoundException("User not found")
+        );
 
         switch (state) {
             case "ALL":
@@ -113,13 +110,13 @@ public class BookingServiceImpl implements BookingService {
     @Override
     public List<Booking> findAllByOwner(long userId, String state) {
         LocalDateTime now = LocalDateTime.now();
-        userValidationService.checkUserIsExist(userId);
-        User user = userRepository.findById(userId).get();
+        User user = userRepository.findById(userId).orElseThrow(
+                () -> new NotFoundException("User not found")
+        );
 
         switch (state) {
             case "ALL":
-                List<Booking> books =  bookingRepository.findAllBookingsByOwner(user);
-                return books;
+                return bookingRepository.findAllBookingsByOwner(user);
             case "CURRENT":
                 return bookingRepository.findCurrentBookingsByOwner(user, now);
             case "PAST":
