+++ src/main/java/ru/practicum/shareit/user/UserController.java
@@ -42,7 +42,7 @@ public class UserController {
      */
     @PatchMapping("/{userId}")
     public User updateUser(@RequestBody UserDto userDto, @PathVariable long userId) {
-        log.info("Updating user id: " + userId);
+        log.info("Updating user id: {}", userId);
         return userService.updateUser(userDto, userId);
     }
 
@@ -53,7 +53,7 @@ public class UserController {
      */
     @GetMapping("/{userId}")
     public User findUserById(@PathVariable long userId) {
-        log.info("Getting user by id: " + userId);
+        log.info("Getting user by id: {}", userId);
         return userService.findUserById(userId);
     }
 
@@ -63,7 +63,7 @@ public class UserController {
      */
     @DeleteMapping("/{userId}")
     public void deleteUser(@PathVariable long userId) {
-        log.info("Deleting user by id: " + userId);
+        log.info("Deleting user by id: {}", userId);
         userService.deleteUser(userId);
     }
 }
