+++ src/main/java/ru/practicum/shareit/util/BookingValidationService.java
@@ -1,19 +1,13 @@
-package ru.practicum.shareit.booking;
+package ru.practicum.shareit.util;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import ru.practicum.shareit.exceptions.NotFoundException;
+import ru.practicum.shareit.booking.BookingDto;
 import ru.practicum.shareit.exceptions.ValidationException;
 
 import java.time.LocalDateTime;
-import java.util.Optional;
 
-@Service
-@RequiredArgsConstructor
 public class BookingValidationService {
-    private final BookingRepository bookingRepository;
 
-    public void validate(BookingDto bookingDto) {
+    public static void validate(BookingDto bookingDto) {
         if (bookingDto.getItemId() == 0 ||
         bookingDto.getStart() == null ||
         bookingDto.getEnd() == null ||
@@ -24,11 +18,5 @@ public class BookingValidationService {
         ) {
             throw new ValidationException("WRONG booking request");
         }
-    }
-
-    public void checkBookingIsExist(long bookingId) {
-        Optional<Booking> booking = bookingRepository.findById(bookingId);
-        if (booking.isEmpty())
-            throw new NotFoundException("Booking not found");
     }
 }
